rename project to rerun_examples
