+++ .devcontainer/devcontainer.json
@@ -1,5 +1,5 @@
 {
-	"name": "python_template",
+	"name": "rerun_examples",
 	"build": {
 		"dockerfile": "Dockerfile",
 		"context": ".."
