+++ docs/intro.rst
@@ -17,11 +17,11 @@ This has basic setup for
 
 ## Continuous Integration Status
 
-[![Ci](https://github.com/blooop/python_template/actions/workflows/ci.yml/badge.svg?branch=main)](https://github.com/blooop/python_template/actions/workflows/ci.yml?query=branch%3Amain)
-[![Codecov](https://codecov.io/gh/blooop/python_template/branch/main/graph/badge.svg?token=Y212GW1PG6)](https://codecov.io/gh/blooop/python_template)
-[![GitHub issues](https://img.shields.io/github/issues/blooop/python_template.svg)](https://GitHub.com/blooop/python_template/issues/)
-[![GitHub pull-requests merged](https://badgen.net/github/merged-prs/blooop/python_template)](https://github.com/blooop/python_template/pulls?q=is%3Amerged)
-[![GitHub release](https://img.shields.io/github/release/blooop/python_template.svg)](https://GitHub.com/blooop/python_template/releases/)
+[![Ci](https://github.com/blooop/rerun_examples/actions/workflows/ci.yml/badge.svg?branch=main)](https://github.com/blooop/rerun_examples/actions/workflows/ci.yml?query=branch%3Amain)
+[![Codecov](https://codecov.io/gh/blooop/rerun_examples/branch/main/graph/badge.svg?token=Y212GW1PG6)](https://codecov.io/gh/blooop/rerun_examples)
+[![GitHub issues](https://img.shields.io/github/issues/blooop/rerun_examples.svg)](https://GitHub.com/blooop/rerun_examples/issues/)
+[![GitHub pull-requests merged](https://badgen.net/github/merged-prs/blooop/rerun_examples)](https://github.com/blooop/rerun_examples/pulls?q=is%3Amerged)
+[![GitHub release](https://img.shields.io/github/release/blooop/rerun_examples.svg)](https://GitHub.com/blooop/rerun_examples/releases/)
 [![License](https://img.shields.io/pypi/l/bencher)](https://opensource.org/license/mit/)
 [![Python](https://img.shields.io/badge/python-3.10%20%7C%203.11-blue)](https://www.python.org/downloads/release/python-310/)
